# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma DTI Níjar — API
  version: "1.0.0"
  description: |
    API REST/JSON de la **Plataforma de Destino Turístico Inteligente (DTI) Smart City**
    del Ayuntamiento de Níjar.

    **Expediente:** 18962/2025
    **Adjudicatario:** IT DIGITTAL
    **Marco:** PRTR — NextGenerationEU — Componente 14
    **Estándares:** UNE 178104 · FIWARE Smart Data Models · ENS Nivel Medio · WCAG 2.1 AA

    ## Versiones y compatibilidad
    - Esta especificación corresponde a la **API v1**.
    - Las rutas están prefijadas con `/api/v1`.
    - Compatibilidad retroactiva garantizada en versiones menores.

    ## Autenticación
    Todos los endpoints (excepto `health`, `ready` y los de autenticación) requieren un
    token JWT válido emitido por `/api/v1/auth/login`. Los tokens incluyen scopes RBAC
    asociados al rol del usuario (administrador_tic, gestor_contenidos, analista_datos,
    operador_smart_office, auditor).
  contact:
    name: IT DIGITTAL
    url: https://www.turismonijar.es
  license:
    name: Propiedad del Ayuntamiento de Níjar (PCAP Anexo XIII)
servers:
  - url: https://api.dti.nijar.es/api/v1
    description: Producción
  - url: https://staging.api.dti.nijar.es/api/v1
    description: Pre-producción
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /tourism/resources:
    get:
      tags: [tourism]
      summary: Listar recursos turísticos
      parameters:
        - { in: query, name: categoria, schema: { type: string } }
        - { in: query, name: municipio, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: page_size, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedRecursos" }
    post:
      tags: [tourism]
      summary: Crear un recurso turístico
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecursoTuristicoIn" }
      responses:
        "201":
          description: Recurso creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecursoTuristicoOut" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /tourism/resources/{resource_id}:
    parameters:
      - { in: path, name: resource_id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [tourism]
      summary: Detalle de un recurso turístico
      responses:
        "200":
          description: Recurso
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecursoTuristicoOut" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [tourism]
      summary: Actualizar un recurso turístico
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecursoTuristicoIn" }
      responses:
        "200":
          description: Recurso actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecursoTuristicoOut" }
    delete:
      tags: [tourism]
      summary: Despublicar un recurso (soft delete)
      responses:
        "204": { description: Recurso despublicado }

  # ---------------- TOURISM: EVENTS ----------------
  /tourism/events:
    get:
      tags: [tourism]
      summary: Listar eventos turísticos
      parameters:
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
        - { in: query, name: tipo, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: page_size, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedEventos" }
  /tourism/events/{event_id}:
    parameters:
      - { in: path, name: event_id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [tourism]
      summary: Detalle de un evento
      responses:
        "200":
          description: Evento
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventoTuristicoOut" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- TOURISM: SERVICES ----------------
  /tourism/services:
    get:
      tags: [tourism]
      summary: Listar servicios turísticos
      parameters:
        - { in: query, name: tipo, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: page_size, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedServicios" }
  /tourism/services/{service_id}:
    parameters:
      - { in: path, name: service_id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [tourism]
      summary: Detalle de un servicio
      responses:
        "200":
          description: Servicio
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServicioOut" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- IOT ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    BadRequest:
      description: Petición mal formada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
...
    GeoPoint:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Point] }
        coordinates:
          type: array
          items: { type: number }
          minItems: 2
          maxItems: 2
    I18nText:
      type: object
      properties:
        es: { type: string }
        en: { type: string }
        de: { type: string }
        fr: { type: string }

    # ---- Recurso Turístico ----
    RecursoTuristicoIn:
      type: object
      required: [urn, nombre, categoria]
      properties:
        urn: { type: string, pattern: "^urn:ngsi-ld:RecursoTuristico:nijar:[a-z0-9-]+$" }
        nombre: { type: string }
        categoria:
          type: string
          enum: [playa, monumento, ruta, mirador, centro_visitantes, parque_natural, museo, yacimiento, punto_interes, oficina_turismo]
        descripcion_corta: { type: string }
        nombre_i18n: { $ref: "#/components/schemas/I18nText" }
        descripcion_i18n: { $ref: "#/components/schemas/I18nText" }
        ubicacion: { $ref: "#/components/schemas/GeoPoint" }
        direccion: { type: string }
        municipio: { type: string, default: Níjar }
        codigo_postal: { type: string }
        telefono: { type: string }
        email: { type: string, format: email }
        web: { type: string, format: uri }
        horario: { type: object }
        accesibilidad: { type: object }
        servicios_disponibles: { type: array, items: { type: string } }
        etiquetas: { type: array, items: { type: string } }
        imagenes: { type: array, items: { type: string, format: uri } }
        publicado: { type: boolean, default: false }
        activo: { type: boolean, default: true }
    RecursoTuristicoOut:
      allOf:
        - { $ref: "#/components/schemas/RecursoTuristicoIn" }
        - type: object
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    PaginatedRecursos:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/RecursoTuristicoOut" }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }

    # ---- Evento ----
    EventoTuristicoOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        urn: { type: string }
        nombre: { type: string }
        tipo: { type: string }
        descripcion: { type: string }
        fecha_inicio: { type: string, format: date-time }
        fecha_fin: { type: string, format: date-time }
        ubicacion: { $ref: "#/components/schemas/GeoPoint" }
        organizador: { type: string }
        publicado: { type: boolean }
    PaginatedEventos:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/EventoTuristicoOut" }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }

    # ---- Servicio ----
    ServicioOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        urn: { type: string }
        nombre: { type: string }
        tipo: { type: string }
        descripcion: { type: string }
        ubicacion: { $ref: "#/components/schemas/GeoPoint" }
        direccion: { type: string }
        municipio: { type: string }
        telefono: { type: string }
        web: { type: string, format: uri }
        rango_precios: { type: string }
        valoracion_media: { type: number }
        publicado: { type: boolean }
    PaginatedServicios:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ServicioOut" }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }

    # ---- Sensor / Observación ----
...